@@ -1,112 +1,119 @@
 {
-  "name": "ember-jsonapi-docs-generator",
-  "version": "3.0.0",
-  "description": "Generates API docs for the official ember projects",
-  "homepage": "https://github.com/ember-learn/ember-jsonapi-docs",
-  "bugs": "https://github.com/ember-learn/ember-jsonapi-docs/issues",
-  "repository": "ember-learn/ember-jsonapi-docs",
-  "license": "MIT",
-  "author": "Sivakumar Kailasam @sivakumar-kailasam",
-  "contributors": [
-    "Sivakumar Kailasam <siva@rubberduck.engineering>",
-    "Todd Jordan",
-    "Ember.JS Learning team"
-  ],
-  "files": [
-    "/bin",
-    "/lib",
-    "/npm-shrinkwrap.json",
-    "/oclif.manifest.json"
-  ],
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "bin": {
-    "ember-docgen": "./bin/run"
-  },
-  "directories": {
-    "test": "test"
-  },
-  "scripts": {
-    "prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
-    "postpack": "rm -f oclif.manifest.json",
-    "test": "nyc --extension .ts mocha --forbid-only \"test/**/*.test.ts\"",
-    "posttest": "tslint -p test -t stylish",
-    "version": "oclif-dev readme && git add README.md"
-  },
-  "dependencies": {
-    "@ember-learn/released-js-docs": "0.0.0",
-    "@oclif/command": "^1",
-    "@oclif/config": "^1",
-    "@oclif/plugin-help": "^2",
-    "command-exists": "^1.2.8",
-    "compare-versions": "^3.5.1",
-    "download": "^7.1.0",
-    "ember-rfc176-data": "^0.3.12",
-    "esm": "^3.2.25",
-    "execa": "^2.0.4",
-    "file-extension": "^4.0.5",
-    "fs-extra": "^8.1.0",
-    "gatsby-remark-vscode": "^1.3.1",
-    "glob": "^7.1.4",
-    "hard-rejection": "^2.1.0",
-    "hast-util-raw": "^5.0.1",
-    "http-server": "^0.11.1",
-    "inflected": "^2.0.4",
-    "keyfinder": "^1.0.0",
-    "lodash": "^4.17.15",
-    "lodash.flatten": "^4.4.0",
-    "lodash.uniq": "^4.5.0",
-    "mdast-util-to-hast": "^7.0.0",
-    "minimist": "^1.2.0",
-    "mkdir-promise": "^1.0.0",
-    "mkdirp": "^0.5.1",
-    "ora": "^3.4.0",
-    "pkg-versions": "^2.1.0",
-    "rehype-stringify": "^6.0.1",
-    "remark-parse": "^7.0.2",
-    "rev-file": "^3.0.0",
-    "rimraf": "^2.6.3",
-    "rsvp": "^4.8.5",
-    "semver-compare": "^1.0.0",
-    "semver-extra": "^3.0.0",
-    "semver-utils": "^1.1.4",
-    "tslib": "^1",
-    "unified": "^8.4.2",
-    "yuidoc-to-jsonapi": "^0.4.1"
-  },
-  "devDependencies": {
-    "@oclif/dev-cli": "^1",
-    "@oclif/test": "^1",
-    "@oclif/tslint": "^3",
-    "@types/chai": "^4",
-    "@types/mocha": "^5",
-    "@types/node": "^10",
-    "chai": "^4.2.0",
-    "chalk": "^2.4.2",
-    "eslint": "6.4.0",
-    "eslint-plugin-node": "10.0.0",
-    "eslint-plugin-prettier": "3.1.1",
-    "globby": "^10",
-    "husky": "^3.0.5",
-    "mocha": "^6.2.0",
-    "nyc": "^14",
-    "prettier": "1.18.2",
-    "pretty-quick": "^1.11.1",
-    "pretty-time": "^1.1.0",
-    "request": "^2.88.0",
-    "ts-node": "^8",
-    "tslint": "^5",
-    "tslint-config-prettier": "^1.18.0",
-    "typescript": "^3.3"
-  },
-  "cacheDirectories": [
-    "tmp"
-  ],
-  "oclif": {
-    "commands": "./lib/commands",
-    "bin": "ember-docgen",
-    "plugins": [
-      "@oclif/plugin-help"
-    ]
-  }
+	"name": "ember-jsonapi-docs-generator",
+	"version": "3.0.0",
+	"description": "Generates API docs for the official ember projects",
+	"homepage": "https://github.com/ember-learn/ember-jsonapi-docs",
+	"bugs": "https://github.com/ember-learn/ember-jsonapi-docs/issues",
+	"repository": "ember-learn/ember-jsonapi-docs",
+	"license": "MIT",
+	"author": "Sivakumar Kailasam @sivakumar-kailasam",
+	"contributors": [
+		"Sivakumar Kailasam <siva@rubberduck.engineering>",
+		"Todd Jordan",
+		"Ember.JS Learning team"
+	],
+	"files": [
+		"/bin",
+		"/lib",
+		"/npm-shrinkwrap.json",
+		"/oclif.manifest.json"
+	],
+	"main": "lib/index.js",
+	"types": "lib/index.d.ts",
+	"bin": {
+		"ember-docgen": "./bin/run"
+	},
+	"directories": {
+		"test": "test"
+	},
+	"scripts": {
+		"prepack": "rm -rf lib && tsc -b && oclif-dev manifest && oclif-dev readme",
+		"postpack": "rm -f oclif.manifest.json",
+		"test": "nyc --extension .ts mocha --forbid-only \"test/**/*.test.ts\"",
+		"posttest": "tslint -p test -t stylish",
+		"version": "oclif-dev readme && git add README.md"
+	},
+	"husky": {
+		"hooks": {
+			"pre-commit": "pretty-quick --staged"
+		}
+	},
+	"dependencies": {
+		"@ember-learn/released-js-docs": "0.0.0",
+		"@oclif/command": "^1",
+		"@oclif/config": "^1",
+		"@oclif/plugin-help": "^2",
+		"command-exists": "^1.2.8",
+		"compare-versions": "^3.5.1",
+		"download": "^7.1.0",
+		"ember-rfc176-data": "^0.3.12",
+		"execa": "^3.2.0",
+		"file-extension": "^4.0.5",
+		"fs-extra": "^8.1.0",
+		"gatsby-remark-vscode": "^1.3.1",
+		"glob": "^7.1.6",
+		"hard-rejection": "^2.1.0",
+		"hast-util-raw": "^5.0.1",
+		"http-server": "^0.11.1",
+		"inflected": "^2.0.4",
+		"keyfinder": "^1.0.0",
+		"lodash": "^4.17.15",
+		"lodash.flatten": "^4.4.0",
+		"lodash.uniq": "^4.5.0",
+		"mdast-util-to-hast": "^7.0.0",
+		"mkdir-promise": "^1.0.0",
+		"mkdirp": "^0.5.1",
+		"ora": "^4.0.2",
+		"pkg-versions": "^2.1.0",
+		"rehype-stringify": "^6.0.1",
+		"remark-parse": "^7.0.2",
+		"rev-file": "^3.0.0",
+		"rimraf": "^3.0.0",
+		"rsvp": "^4.8.5",
+		"semver-compare": "^1.0.0",
+		"semver-extra": "^3.0.0",
+		"semver-utils": "^1.1.4",
+		"tslib": "^1",
+		"unified": "^8.4.2",
+		"yuidoc-to-jsonapi": "^0.4.1"
+	},
+	"devDependencies": {
+		"@oclif/dev-cli": "^1",
+		"@oclif/test": "^1",
+		"@oclif/tslint": "^3",
+		"@types/chai": "^4",
+		"@types/fs-extra": "^8.0.1",
+		"@types/lodash": "^4.14.146",
+		"@types/mocha": "^5",
+		"@types/node": "^12",
+		"@types/pretty-time": "^1.1.0",
+		"@types/rimraf": "^2.0.3",
+		"@types/rsvp": "^4.0.3",
+		"chai": "^4.2.0",
+		"eslint": "6.6.0",
+		"eslint-plugin-node": "10.0.0",
+		"eslint-plugin-prettier": "3.1.1",
+		"globby": "^10",
+		"husky": "^3.0.9",
+		"mocha": "^6.2.2",
+		"nyc": "^14",
+		"prettier": "1.19.1",
+		"pretty-quick": "^2.0.1",
+		"pretty-time": "^1.1.0",
+		"request": "^2.88.0",
+		"ts-node": "^8",
+		"tslint": "^5",
+		"tslint-config-prettier": "^1.18.0",
+		"typescript": "^3.7"
+	},
+	"cacheDirectories": [
+		"tmp"
+	],
+	"oclif": {
+		"commands": "./lib/commands",
+		"bin": "ember-docgen",
+		"plugins": [
+			"@oclif/plugin-help"
+		]
+	}
 }
