@@ -1,4 +1,4 @@
-import fs from 'fs-extra'
+import * as fs from 'fs-extra'
 import saveDoc from './save-document'
 
 export async function filler1(projectName, docs) {
