@@ -1,4 +1,4 @@
-import _ from 'lodash'
+import * as _ from 'lodash'
 import { Promise } from 'rsvp'
 import { byType } from './filter-jsonapi-doc'
 
