@@ -1,6 +1,6 @@
-import fs from 'fs-extra'
-import glob from 'glob'
-import path from 'path'
+import * as fs from 'fs-extra'
+import * as glob from 'glob'
+import * as path from 'path'
 
 export default function readDocs(projects, specificVersion = '') {
 	return projects.reduce(async (prevPromise, projectName) => {
