@@ -1,4 +1,4 @@
-import _ from 'lodash'
+import * as _ from 'lodash'
 
 function generateID(type, projectName, version, id) {
 	if (!projectName) {
